requirements.txt
streamlit
pandas
numpy
scikit-learn
pillow
opencv-python
matplotlib
seaborn
plotly
